+++ src/components/Footer/index.js
@@ -19,7 +19,7 @@ export class Footer extends React.Component {
           <div className="col-md-3">
             <ul className="list-unstyled">
               <li><a href="./home">Home</a></li>
-              <li><a href="./contact">Contact</a></li>
+              <li><a href="./share">Share</a></li>
               <li><a href="./motive">Motive</a></li>
             </ul>
           </div>